Conflicts:
	app/post/postService.js

+++ app/post/postDetailController.js
@@ -5,22 +5,22 @@
     angular.module('MyApp')
         .controller('postDetailController', postDetailController);
 
-    postDetailController.$inject = ['$scope', 'postService', '$location', '$routeParams', 'headerService', '$window'];
+    postDetailController.$inject = ['$scope', 'ngDialog', 'postService', '$location', '$routeParams', 'headerService', '$window'];
 
-    function postDetailController($scope, postService, $location, $routeParam, headerService, $window) {
+    function postDetailController($scope, ngDialog, postService, $location, $routeParam, headerService, $window) {
         $scope.iconLike = false;
         $scope.iconFavorite = false;
         $scope.editing = false;
         $scope.editingComment = false;
-
         postService.getPost($routeParam.id)
-            .success(function(data){
+            .success(function (data) {
                 $scope.post = data.post;
+
                 $scope.post.created_at = beautyDate.prettyDate($scope.post.created_at);
-                for (var i=0; i< $scope.post.tag_picture.length; i++){
+                for (var i = 0; i < $scope.post.tag_picture.length; i++) {
                     $scope.post.tag_picture[i].top = $scope.post.tag_picture[i].top - 20;
                 }
-                for (var i=0; i< $scope.post.comments.length; i++){
+                for (var i = 0; i < $scope.post.comments.length; i++) {
                     $scope.post.comments[i].created_at = beautyDate.prettyDate($scope.post.comments[i].created_at);
                 }
                 headerService.loginUser()
@@ -52,29 +52,41 @@
                         event.preventDefault();
                         headerService.openLogin();
                     } else {
-                        var data = {
-                            'content': this.comment,
-                            'type_comment': 0,
-                            'type_id': $routeParam.id,
-                            'user_id': $scope.loginUser.id
-                        };
-                        this.comment = null;
-                        //$scope.post.comments.push({
-                        //    'content': data.content,
-                        //    'created_at': 'Just now',
-                        //    'user': {
-                        //        'username': $scope.loginUser.username,
-                        //        'id': $scope.loginUser.id,
-                        //        'picture_profile': $scope.loginUser.picture_profile
-                        //    }
-                        //});
+                        if (this.comment != null) {
+                            if (this.comment.length > 255) {
+                                ngDialog.open({
+                                    template: 'app/post/templates/alertInpTxtLength.html',
+                                    className: 'ngdialog-theme-plain-custom',
 
-                        postService.saveComment(data)
-                            .success(function (data) {
-                            })
-                            .error(function (data) {
+                                    controller: ['$scope', 'postService', function ($scope, postService) {
+                                    }]
+                                });
+                            } else {
+                                var data = {
+                                    'content': this.comment,
+                                    'type_comment': 0,
+                                    'type_id': $routeParam.id,
+                                    'user_id': $scope.loginUser.id
+                                };
+                                this.comment = null;
+                                $scope.post.comments.push({
+                                    'content': data.content,
+                                    'created_at': 'Just now',
+                                    'user': {
+                                        'username': $scope.loginUser.username,
+                                        'id': $scope.loginUser.id,
+                                        'picture_profile': $scope.loginUser.picture_profile
+                                    }
+                                });
 
-                            });
+                                postService.saveComment(data)
+                                    .success(function (data) {
+                                    })
+                                    .error(function (data) {
+
+                                    });
+                            }
+                        }
                     }
                 };
                 $scope.likeOrDislike = function () {
@@ -141,25 +153,47 @@
                     $scope.editContent = content;
                 };
 
-                $scope.submitEditComment = function (index) {
-                    $scope.post.comments[index].content = this.editContent;
-                    $scope.post.comments[index].editing = null;
-                    postService.editPostComment({
-                        id: $scope.post.comments[index].id,
-                        content: this.editContent
-                    }).success(function (data) {
+                $scope.submitEditComment = function (index, editContent) {
+                    if (editContent.length > 255) {
+                        ngDialog.open({
+                            template: 'app/post/templates/alertInpTxtLength.html',
+                            className: 'ngdialog-theme-plain-custom',
 
-                    }).error();
+                            controller: ['$scope', 'postService', function ($scope, postService) {
+                            }]
+                        });
+                    } else {
+                        $scope.post.comments[index].content = this.editContent;
+                        $scope.post.comments[index].editing = null;
+                        postService.editPostComment({
+                            id: $scope.post.comments[index].id,
+                            content: this.editContent
+                        }).success(function (data) {
+
+                        }).error();
+                    }
                 };
 
-                $scope.deleteCommentIndex = function (index) {
+                $scope.deleteCommentIndex = function (index, comments) {
+                    ngDialog.open({
+                        template: 'app/post/templates/confirmDeletePost.html',
+                        className: 'ngdialog-theme-plain-custom',
 
-                    postService.deletePostComment({
-                        id: $scope.post.comments[index].id
-                    }).success(function (data) {
+                        controller: ['$scope', function ($scope) {
+                            $scope.close = function () {
+                                ngDialog.close();
+                            };
+                            $scope.confirm = function () {
+                                ngDialog.close();
+                                postService.deletePostComment({
+                                    id: comments[index].id
+                                }).success(function (data) {
 
-                    }).error();
-                    $scope.post.comments.splice(index, 1);
+                                }).error();
+                                comments.splice(index, 1);
+                            }
+                        }]
+                    });
                 };
 
                 $scope.closeEditComment = function (index) {
@@ -189,11 +223,24 @@
                 };
 
                 $scope.deletePostInside = function () {
-                    postService.delete({
-                        id: $routeParam.id
-                    }).success(function (data) {
-                        $window.location.reload();
-                    }).error();
+                    ngDialog.open({
+                        template: 'app/post/templates/confirmDeletePost.html',
+                        className: 'ngdialog-theme-plain-custom',
+
+                        controller: ['$scope', function ($scope) {
+                            $scope.close = function () {
+                                ngDialog.close();
+                            };
+                            $scope.confirm = function () {
+                                ngDialog.close();
+                                postService.delete({
+                                    id: $routeParam.id
+                                }).success(function (data) {
+                                    $window.location.reload();
+                                }).error();
+                            }
+                        }]
+                    });
                 };
 
                 $scope.closeDialog = function () {
